requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn



# ---------------------------------
# PyTorch and Related Libraries

torch
torchvision
torchaudio

# -------------------------------

# ---------------------------------
# required for Hugging Face Embeddings

sentence-transformers

# -------------------------------

langchain-community
bs4
faiss-cpu 
pypdf 
arxiv
pymupdf
wikipedia
lxml
langchain-groq
langgraph
langgraph-cli[inmem]
